PELE-462 Moved coords extraction to another module
