+++ pele_platform/PPI/main.py
@@ -1,6 +1,5 @@
 import os
-import glob
-from pele_platform.site_finder.cluster import cluster_best_structures
+from pele_platform.PPI.cluster import cluster_best_structures
 from pele_platform.PPI.simulation_launcher import launch_simulation
 from pele_platform.PPI.preparation import prepare_structure
 from pele_platform.Utilities.Helpers.helpers import cd, is_repeated, is_last
